pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["venv", "logs", "test_data", "tests"]

[project]
name = "elec_price_monitor"
version = "0.1.2"
description = "Monitor electricity prices and communicate with the uC"
authors = [
  {name = "Your Name", email = "you@example.com" },
]

dependencies = [
  "Pillow",
  "Requests",
  "matplotlib"
]

[project.scripts]
#elec_price_monitor-cli = "elec_price_monitor:main"

